tooling: Add feature "uefi" and use runner to execute
The first change is that we put a feature flag "uefi" into
the Cargo.toml. With that feature flag we can specify whether
we are compiling for an UEFI environment or a bare metal environment.
Depending on that we include boot.rs which includes boot.s
( _start() entry point for bare metal in boot.s).
Or we include efi.rs (efi_main()) in case the
"uefi" feature flag is set.

The second change is that we are using the runner field
to run our compiled binary in qemu.
With this we can simply do a 'cargo run --target $TARGET'
and it will compile the binary and start qemu for us.