driver: qemu ramfb driver
On embedded platforms like aarch64 or risc-v there
is no default display output. A simple way to get
a display on those platforms on QEMU is to use the
ramfb device.
This commit adds a driver for the ramfb device that
initializes it and then lets you draw to the
user provided framebuffer.